requirements.txt
absl-py==0.9.0
backcall==0.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
future==0.18.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
grpcio==1.27.2
idna==2.8
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
joblib==0.14.1
Markdown==3.2.1
numpy==1.18.1
oauthlib==3.1.0
ogb==1.0.1
pandas==1.0.1
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pymongo==3.10.1
python-dateutil==2.8.1
pytorch-lightning==0.7.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
tensorboard==2.1.0
test-tube==0.7.5
torch==1.4.0
torchvision==0.4.2
tqdm==4.42.1
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
